Separates into different files
I'm now serving the game from a GithubPages url, so that's easy.

That means that I can load modules using javascript import syntax
(the thing that was preventing this before is that I wanted to be able to hand the game to someone & they could immediately play it. If you try to import other files, browsers get upset about CORS, but if it's a server that works fine. Because it's easy to host now on github pages, I'm happy with making it a real server.)