Fix bug: grow some 0s
I had put -1s in the map where I wanted the player to be unable to go.

However, default behavior is that if the map doesn't recognize the thing there, you can grow it.
This replaces those -1s with 9s, which were once used to block player movement.

Since 9s should now function identically to being off the map, I have them share the "0" character as the display value if you're off the map.